+++ UnitTests/BraintreeDataCollectorTests/BTDataCollector_Tests.swift
@@ -5,6 +5,8 @@ import PPRiskMagnes
 @testable import BraintreeDataCollector
 
 class BTDataCollector_Tests: XCTestCase {
+    
+    let authorization: String = "sandbox_9dbg82cq_dcpspy2brwdjr3qn"
 
     func testCollectDeviceData_collectsAllData() {
         let config: [String: Any] = ["environment": "development"]
@@ -12,7 +14,9 @@ class BTDataCollector_Tests: XCTestCase {
         let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: config)
         
-        let dataCollector = BTDataCollector(apiClient: mockAPIClient)
+        let dataCollector = BTDataCollector(authorization: authorization)
+        dataCollector.apiClient = mockAPIClient
+        
         let expectation = self.expectation(description: "Returns fraud data")
         dataCollector.collectDeviceData { deviceData, _ in
             if let deviceData = deviceData {
@@ -39,7 +43,9 @@ class BTDataCollector_Tests: XCTestCase {
         let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: config)
 
-        let dataCollector = BTDataCollector(apiClient: mockAPIClient)
+        let dataCollector = BTDataCollector(authorization: authorization)
+        dataCollector.apiClient = mockAPIClient
+        
         let expectation = self.expectation(description: "Returns fraud data")
 
         dataCollector.collectDeviceData { deviceData, _ in
@@ -61,7 +67,8 @@ class BTDataCollector_Tests: XCTestCase {
         let configuration = BTConfiguration(json: BTJSON(value: config))
         let pairingID = "random pairing id"
         let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
-        let dataCollector = BTDataCollector(apiClient: mockAPIClient)
+        let dataCollector = BTDataCollector(authorization: authorization)
+        dataCollector.apiClient = mockAPIClient
 
         XCTAssertEqual(pairingID, dataCollector.clientMetadataID(pairingID))
         XCTAssertNotEqual(pairingID, dataCollector.generateClientMetadataID(with: configuration))
@@ -75,7 +82,8 @@ class BTDataCollector_Tests: XCTestCase {
         
         let configuration = BTConfiguration(json: BTJSON(value: config))
         let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
-        let dataCollector = BTDataCollector(apiClient: mockAPIClient)
+        let dataCollector = BTDataCollector(authorization: authorization)
+        dataCollector.apiClient = mockAPIClient
 
         let clientMetaDataID = dataCollector.generateClientMetadataID(with: configuration)
         let clientMetaDataID2 = dataCollector.clientMetadataID("some pairing id")
@@ -89,7 +97,8 @@ class BTDataCollector_Tests: XCTestCase {
         
         let configuration = BTConfiguration(json: BTJSON(value: config))
         let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
-        let dataCollector = BTDataCollector(apiClient: mockAPIClient)
+        let dataCollector = BTDataCollector(authorization: authorization)
+        dataCollector.apiClient = mockAPIClient
 
         XCTAssertNotNil(dataCollector.generateClientMetadataID(with: configuration))
         XCTAssertEqual(dataCollector.getMagnesEnvironment(from: dataCollector.config), MagnesSDK.Environment.SANDBOX)
@@ -102,7 +111,8 @@ class BTDataCollector_Tests: XCTestCase {
         
         let configuration = BTConfiguration(json: BTJSON(value: config))
         let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
-        let dataCollector = BTDataCollector(apiClient: mockAPIClient)
+        let dataCollector = BTDataCollector(authorization: authorization)
+        dataCollector.apiClient = mockAPIClient
 
         XCTAssertNotNil(dataCollector.generateClientMetadataID(with: configuration))
         XCTAssertEqual(dataCollector.getMagnesEnvironment(from: dataCollector.config), MagnesSDK.Environment.LIVE)
@@ -115,7 +125,8 @@ class BTDataCollector_Tests: XCTestCase {
         
         let configuration = BTConfiguration(json: BTJSON(value: config))
         let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
-        let dataCollector = BTDataCollector(apiClient: mockAPIClient)
+        let dataCollector = BTDataCollector(authorization: authorization)
+        dataCollector.apiClient = mockAPIClient
 
         XCTAssertNotNil(dataCollector.generateClientMetadataID(with: configuration))
         XCTAssertEqual(dataCollector.getMagnesEnvironment(from: dataCollector.config), MagnesSDK.Environment.LIVE)
@@ -123,7 +134,8 @@ class BTDataCollector_Tests: XCTestCase {
     
     func testCollectDeviceData_fetchConfigurationReturnsError_returnError() {
         let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
-        let mockDataCollector = MockBTDataCollector(apiClient: mockAPIClient)
+        let mockDataCollector = MockBTDataCollector(authorization: authorization)
+        mockDataCollector.apiClient = mockAPIClient
 
         mockDataCollector.cannedDataCollectorError = NSError(domain: "FakeConfigError", code: 1, userInfo: [NSLocalizedDescriptionKey:"Fake description"])
 
@@ -144,7 +156,8 @@ class BTDataCollector_Tests: XCTestCase {
     
     func testCollectDeviceData_invalidJSON_returnError() {
         let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
-        let mockDataCollector = MockBTDataCollector(apiClient: mockAPIClient)
+        let mockDataCollector = MockBTDataCollector(authorization: authorization)
+        mockDataCollector.apiClient = mockAPIClient
         
         mockDataCollector.cannedDataCollectorError = BTDataCollectorError.jsonSerializationFailure
 
@@ -165,7 +178,8 @@ class BTDataCollector_Tests: XCTestCase {
     
     func testCollectDeviceData_encodingError_returnError() {
         let mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
-        let mockDataCollector = MockBTDataCollector(apiClient: mockAPIClient)
+        let mockDataCollector = MockBTDataCollector(authorization: authorization)
+        mockDataCollector.apiClient = mockAPIClient
         
         mockDataCollector.cannedDataCollectorError = BTDataCollectorError.encodingFailure