+++ CHANGELOG.md
@@ -35,6 +35,8 @@
     * Update initializer to `BTAmericanExpressClient(authorization:)`
   * BraintreeApplePay
     * Update initializer to `BTApplePayClient(authorization:)`
+  * BraintreeDataCollector
+    * Update initializer to `BTDataCollector(authorization:)`
 
 ## 6.29.0 (2025-02-24)
 * BraintreePayPal